Сверстана кастомная ячейка и она отображается
