+++ ImageFeed.xcodeproj/xcuserdata/a.chuprynenko.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "589F9F88-0516-47F9-8F2C-5D0DC959A691"
+   type = "1"
+   version = "2.0">
+</Bucket>
